refactor channeling signature
